requirements.txt
streamlit==1.41.0
pandas==1.5.3
numpy==1.23.5
yfinance==0.2.18
tensorflow==2.12.0
transformers==4.30.2
scikit-learn==1.2.2
plotly==5.15.0
scipy==1.10.1
newsapi-python==0.2.7
tensorflow-hub==0.12.0
tensorflow-text==2.12.1
dotenv
4a41cc4a79dd438e8faef855e7e62ef7